User Authentication Finilized
Many changes here so if read, read carefully.
User Authentication is now finilized. We can check if a user is logged in before they access anything they cant and redirect them.

Layout has been removed from all pages considering it isnt needed on everything.

AuthLayout has been added to validate that a user is logged in before they can reach that page.

Middleware has been added, so use that if you want to make sure that a user is logged in through the server as well.